+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 ### ✨ Features and improvements
 
 - Remove cooperative gesture screen from the accessibility tree since screenreaders cannot interact with the map using gestures
-- _...Add new stuff here..._
 - Add `cooperated gestures` example to the doc.([#2860](https://github.com/maplibre/maplibre-gl-js/pull/2860))
+- _...Add new stuff here..._
 
 ### 🐞 Bug fixes
 
+- Incorrect distance field of view calculation for negative elevation, fixed by storing min elevation for the tile in view ([#1655](https://github.com/maplibre/maplibre-gl-js/issues/1655), [#2858](https://github.com/maplibre/maplibre-gl-js/pull/2858))
 - Fix reloadCallback not firing on VectorTileWorkerSource.reloadTile ([#1874](https://github.com/maplibre/maplibre-gl-js/pull/1874))
+- _...Add new stuff here..._
 
 ## 3.2.0